requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
PyGithub==2.1.1
pydantic==2.6.1
httpx==0.26.0
pytest==8.0.0
openai==1.12.0
python-multipart==0.0.9
aiohttp==3.9.3 